// src/app/core/enums/MakerCheckerEnum.ts
export enum MakerCheckerEnum
    {
        Reply = 1,
        ReplyClose = 2,
        ReplyEscalate = 3,
        ReplyHold = 4,
        ReplyAwaitingResponse = 5,
        Escalate = 6,
        DirectClose = 7,
        OnHoldAgent = 8,
        CustomerInfoAwaited = 9,
        Close = 10,
        CaseAttach = 11
    }