// src/app/core/interfaces/BulkTicketActions.ts
export interface BulkActionButtons {
    btnbulkdirectclose?: boolean;
    btnbulkreplyapproved?: boolean;
    btnbulkreplyrejected?: boolean;
    btnbulkreopen?: boolean;
    btnbulkassign?: boolean;
    btnbulkescalae?: boolean;
    btnbulkonhold?: boolean;
    btnbulkreject?: boolean;
    btnbulkapprove?: boolean;
    btnbulkreply?: boolean;
}